Revert "config_detail"
