+++ resources/js/routes.js
@@ -6,7 +6,7 @@ import Watch from "./view/pages/Watch";
 import Watchs from "./view/Tabs/WatchTab/Watchs";
 import Musics from "./view/Tabs/WatchTab/Musics";
 import Videos from "./view/Tabs/WatchTab/Videos";
-import Detail from "./view/pages/Detail";
+import VideoDetail from "./view/pages/VideoDetail";
 // auth
 import Login from "./view/pages/Login";
 import Register from "./view/pages/Register";
@@ -17,9 +17,7 @@ import FriendList from "./view/Tabs/FriendTab/FriendList";
 import NewFriend from "./view/Tabs/FriendTab/NewFriend";
 // profile
 import Call from "./view/pages/Call";
-import profile from "./view/components/ProfileComponent/profile";
-// detail
-import VideoDetail from "./view/components/PostComponent/VideoDetailComponent";
+import profile from "./view/components/ProfileComponent/profile"
 const routes = [
     {
         path: "/",
@@ -87,8 +85,8 @@ const routes = [
         // redirect: "/activity",
         children: [
             {
-                path: "",
-                component: profile,
+             path:"",
+             component:profile
             },
             {
                 path: "friendships",
@@ -109,14 +107,8 @@ const routes = [
         meta: { AppProduct: true },
     },
     {
-        path: "/detail",
-        component: Detail,
-        children: [
-            {
-                path: "video",
-                component: VideoDetail,
-            },
-        ],
+        path: "/video-detail",
+        component: VideoDetail,
         meta: { AppProduct: true },
     },
 ];
